ci: Update test workflow to support running end-to-end tests
End-to-end tests interact with the Equinix Metal API and thus require
an API token.  This adds a step to the existing `test` GitHub Actions
workflow that sets a `METAL_AUTH_TOKEN` environment variable that can
be used by end-to-end tests.  The existing test step is updated to set
the `short` flag.  The intent is that we will use the `short` flag to
skip end-to-end tests as documented here:

https://pkg.go.dev/testing#hdr-Skipping