chore: upgrade to Go 1.22 (#461)
As part of this upgrade, I also upgrade golangci-lint to close #406. In
order to successfully install a newer golangci-lint locally, I had to
rip out the `hack/tools` workflow and run the linter in a container
instead. Per the golangci-lint docs, installing binaries or using a
container is preferred for local installation, and `go install` (which
is what the `hack/tools` workflow used) is not guaranteed to work:
https://golangci-lint.run/welcome/install/#install-from-source